rvjs-refactor: change some code
